style: 버튼 네비게이션 link 방식으로 변경
