+++ src/widgets/bottom-navigation/model/routes.ts
